+++ src/core/date.ts
@@ -1,7 +1,7 @@
-import type { RVDate } from '@/types/date';
+import type { RVDate } from "@/types/date"
 
 export function date(): RVDate {
-	return {
-		type: 'date',
-	} as never;
+  return {
+    type: "date",
+  } as never
 }