+++ src/core/or.ts
@@ -1,9 +1,12 @@
-import type { RVOr } from '@/types/or';
-import type { RVSchema } from '@/types/schema';
+import type { RVOr } from "@/types/or"
+import type { RVSchema } from "@/types/schema"
 
-export function or<T extends RVSchema[]>(...schemas: [...T]): RVOr<T> {
-	return {
-		type: 'or',
-		schemas,
-	} as never;
+export function or<A extends RVSchema, B extends RVSchema>(
+  a: A,
+  b: B,
+): RVOr<[A, B]> {
+  return {
+    type: "or",
+    schemas: [a, b],
+  } as never
 }