+++ src/resolvers/typebox/date.ts
@@ -1,6 +1,5 @@
-import { Type } from '@sinclair/typebox';
-
-import type { TDate } from '@sinclair/typebox';
+import type { TDate } from "@sinclair/typebox";
+import { Type } from "@sinclair/typebox";
 
 export function TBDate(): TDate {
 	return Type.Date();