+++ src/resolvers/typebox/required.ts
@@ -1,28 +1,26 @@
-import { Type } from '@sinclair/typebox';
+import { Type } from "@sinclair/typebox"
 
-import type { RVRequiredParams } from '@/types/required';
-import type { SchemaOptions, TSchema } from '@sinclair/typebox';
+import type { SchemaOptions, TSchema } from "@sinclair/typebox"
+import type { RVRequiredParams } from "@/types/required"
 
 export function TBRequired(
-	schema: TSchema,
-	params?: RVRequiredParams,
+  schema: TSchema,
+  params?: RVRequiredParams,
 ): TSchema {
-	const typeBoxParams: SchemaOptions = {};
+  const typeBoxParams: SchemaOptions = {}
 
-	const keys: { [key in keyof RVRequiredParams]: string } = {
-		description: 'description',
-	};
+  const keys: { [key in keyof RVRequiredParams]: string } = {
+    description: "description",
+  }
 
-	if (params) {
-		for (const key in keys) {
-			const mappedKey = keys[
-				key as keyof typeof keys
-			] as keyof RVRequiredParams;
-			if (params[key as keyof RVRequiredParams] !== undefined) {
-				typeBoxParams[mappedKey] = params[key as keyof RVRequiredParams];
-			}
-		}
-	}
+  if (params) {
+    for (const key in keys) {
+      const mappedKey = keys[key as keyof typeof keys] as keyof RVRequiredParams
+      if (params[key as keyof RVRequiredParams] !== undefined) {
+        typeBoxParams[mappedKey] = params[key as keyof RVRequiredParams]
+      }
+    }
+  }
 
-	return Type.Required(schema as unknown as TSchema, typeBoxParams);
+  return Type.Required(schema as unknown as TSchema, typeBoxParams)
 }