+++ src/resolvers/zod/date.ts
@@ -1,0 +1,5 @@
+import z, { type ZodDate } from "zod"
+
+export function ZDDate(): ZodDate {
+  return z.date()
+}