+++ src/types/nullable.d.ts
@@ -1,0 +1,7 @@
+import type { RVSchema } from "./schema"
+
+export interface RVNullable<T extends RVSchema> extends RVSchema {
+  type: "nullable"
+  schema: T
+  infer: T extends { infer: infer U } ? U | null : null
+}